requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-cognitiveservices-speech==1.37.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.32.0
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
pygame==2.5.2
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.0.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
